Got c++ version of software to work in terms of generating a point cloud. Still can be significantly optimized.

+++ include/Coordinate.hpp
@@ -25,9 +25,9 @@ namespace tiltingLIDARScanner {
 
 struct Coordinate
 {
-    double x;
-    double y;
-    double z;
+    float x;
+    float y;
+    float z;
 };
 
 } // namespace tiltingLIDARScanner
